types: Marker's `index` is now `source`
The index in YamlMarker turned out to not be quite what we wanted
so instead of removing the entry entirely, we have transformed it
into a source index instead.  This means that when parsing many
pieces of YAML, the markers will also indicate where the YAML
came from which could be of use.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>