types: Add Deref and DerefMut to the sequence and mapping nodes
Sequences and mappings may need to be changed at runtime (whereas
scalars are considered immutable) and as such we need to access
their underlying containers.  Since we can do that trivially with
the Deref and DerefMut traits, use those.  We don't mind leaking
that our storage is Vec and LinkedHashMap respectively.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>